refactor: move adj reader
